Section 17: Optimizations

+++ API/SignalR/MessageHub.cs
@@ -22,16 +22,19 @@ namespace API.SignalR
 
             var groupName = getGroupName(Context.User.GetUsername(), otherUser);
             await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-            await AddToGroup(groupName);
+            var group = await AddToGroup(groupName);
+
+            await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
             var messages = await messageRepository.GetMessageThread(Context.User.GetUsername(), otherUser!);
 
-            await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+            await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
         }
 
         public override async Task OnDisconnectedAsync(Exception? exception)
         {
-            await RemoveFromMessageGroup();
+            var group = await RemoveFromMessageGroup();
+            await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
             await base.OnDisconnectedAsync(exception);
         }
 
@@ -91,7 +94,7 @@ namespace API.SignalR
             }
         }
 
-        private async Task<bool> AddToGroup(string groupName)
+        private async Task<Group> AddToGroup(string groupName)
         {
             var username = Context.User?.GetUsername() ?? throw new Exception("Could not get user");
 
@@ -106,18 +109,28 @@ namespace API.SignalR
 
             group.Connections.Add(connection);
 
-            return await messageRepository.SaveAllAsync();
+            if(await messageRepository.SaveAllAsync())
+            {
+                return group;
+            }
+
+            throw new HubException("Failed to join group");
         }
 
-        private async Task RemoveFromMessageGroup()
+        private async Task<Group> RemoveFromMessageGroup()
         {
-            var connection = await messageRepository.GetConnection(Context.ConnectionId);
+            var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+            var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
 
-            if (connection != null)
+            if (connection != null && group != null)
             {
                 messageRepository.RemoveConnection(connection);
-                await messageRepository.SaveAllAsync();
+                if (await messageRepository.SaveAllAsync())
+                {
+                    return group;
+                }
             }
+            throw new HubException("Failed to remove from group");
         }
 
         private string getGroupName(string caller, string? other) {
